Working code for Gazebo
This code can run husky_gazebo for simulating Husky in a number of outdoor enviroments.This also include simulating Hesai Pandar 40 LiDAR. It is able to construct the map using Octomap or SLAM LOAM. The user can control Husky using keyboard. Visualization is available for both the instant (single time-frame) pointcloud and the constructed map.